refactor(budget): separate shared modules into shared dir, rename budget-main
