+++ apps/web/src/services/cost-explorer/budget/budget-create/BudgetCreatePage.vue
@@ -15,7 +15,7 @@ import {
 
 import { PHeading } from '@spaceone/design-system';
 
-import BudgetForm from '@/services/cost-explorer/budget/modules/budget-form/BudgetForm.vue';
+import BudgetForm from '@/services/cost-explorer/budget/shared/budget-form/BudgetForm.vue';
 
 export default {
     name: 'BudgetCreatePage',
