+++ apps/web/src/services/cost-explorer/budget/budget-main/BudgetMainPage.vue
@@ -13,7 +13,7 @@ import { i18n } from '@/translations';
 
 import { useManagePermissionState } from '@/common/composables/page-manage-permission';
 
-import BudgetList from '@/services/cost-explorer/budget/modules/budget-list/BudgetList.vue';
+import BudgetList from '@/services/cost-explorer/budget/budget-main/modules/budget-list/BudgetList.vue';
 import { COST_EXPLORER_ROUTE } from '@/services/cost-explorer/route-config';
 
 const currentRoute = SpaceRouter.router.currentRoute;
